--- a/WORKSPACE
+++ b/WORKSPACE
@@ -1,4 +1,4 @@
-workspace(name = "com_github_lestrrat_go_jwx_v2")
+workspace(name = "com_github_lestrrat_go_jwx_v3")
 
 load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")