--- a/format.go
+++ b/format.go
@@ -3,6 +3,8 @@ package jwx
 import (
 	"bytes"
 	"encoding/json"
+
+	"github.com/lestrrat-go/jwx/v3/internal/tokens"
 )
 
 type FormatKind int
@@ -60,9 +62,9 @@ func GuessFormat(payload []byte) FormatKind {
 		return UnknownFormat
 	}
 
-	if payload[0] != '{' {
+	if payload[0] != tokens.OpenCurlyBracket {
 		// Compact format. It's probably a JWS or JWE
-		sep := []byte{'.'} // I want to const this :/
+		sep := []byte{tokens.Period} // I want to const this :/
 
 		// Note: this counts the number of occurrences of the
 		// separator, but the RFC talks about the number of segments.