--- a/go.mod
+++ b/go.mod
@@ -7,9 +7,9 @@ toolchain go1.23.6
 require (
 	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.4.0
 	github.com/goccy/go-json v0.10.3
-	github.com/lestrrat-go/blackmagic v1.0.3
+	github.com/lestrrat-go/blackmagic v1.0.4
 	github.com/lestrrat-go/httprc/v3 v3.0.0
-	github.com/lestrrat-go/option v1.0.1
+	github.com/lestrrat-go/option/v2 v2.0.0
 	github.com/segmentio/asm v1.2.0
 	github.com/stretchr/testify v1.10.0
 	golang.org/x/crypto v0.38.0
@@ -18,6 +18,7 @@ require (
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/lestrrat-go/httpcc v1.0.1 // indirect
+	github.com/lestrrat-go/option v1.0.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/sys v0.33.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect