--- a/internal/base64/base64.go
+++ b/internal/base64/base64.go
@@ -16,6 +16,8 @@ type Encoder interface {
 	Encode([]byte, []byte)
 	EncodedLen(int) int
 	EncodeToString([]byte) string
+
+	AppendEncode([]byte, []byte) []byte
 }
 
 var muEncoder sync.RWMutex
@@ -60,6 +62,10 @@ func Encode(src []byte) []byte {
 
 func EncodeToString(src []byte) string {
 	return getEncoder().EncodeToString(src)
+}
+
+func AppendEncode(dst []byte, src []byte) []byte {
+	return getEncoder().AppendEncode(dst, src)
 }
 
 func EncodeUint64ToString(v uint64) string {