--- a/internal/jwxtest/BUILD.bazel
+++ b/internal/jwxtest/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
     importpath = "github.com/lestrrat-go/jwx/v3/internal/jwxtest",
     visibility = ["//:__subpackages__"],
     deps = [
+        "//internal/tokens",
         "//jwa",
         "//jwe",
         "//jwk",