--- a/internal/pool/big_int.go
+++ b/internal/pool/big_int.go
@@ -15,7 +15,6 @@ func destroyBigInt(i *big.Int) {
 }
 
 func GetBigInt() *big.Int {
-	//nolint:forcetypeassert
 	return bigIntPool.Get()
 }