--- a/internal/pool/big_int.go
+++ b/internal/pool/big_int.go
@@ -6,18 +6,14 @@ import (
 
 var bigIntPool = New(allocBigInt, destroyBigInt)
 
+func BigInt() *Pool[*big.Int] {
+	return bigIntPool
+}
+
 func allocBigInt() interface{} {
 	return &big.Int{}
 }
 
 func destroyBigInt(i *big.Int) {
 	i.SetInt64(0)
-}
-
-func GetBigInt() *big.Int {
-	return bigIntPool.Get()
-}
-
-func ReleaseBigInt(i *big.Int) {
-	bigIntPool.Put(i)
 }