--- a/internal/pool/byte_slice.go
+++ b/internal/pool/byte_slice.go
@@ -3,7 +3,6 @@ package pool
 var bytesSlicePool = New(allocByteSlice, destroyByteSlice)
 
 func GetByteSlice() *[]byte {
-	//nolint:forcetypeassert
 	return bytesSlicePool.Get()
 }