--- a/internal/pool/string_slice.go
+++ b/internal/pool/string_slice.go
@@ -12,7 +12,6 @@ func destroyStringSlice(s *[]string) {
 }
 
 func GetStringSlice() *[]string {
-	//nolint:forcetypeassert
 	return stringSlicePool.Get()
 }