--- a/internal/pool/string_slice.go
+++ b/internal/pool/string_slice.go
@@ -2,6 +2,10 @@ package pool
 
 var stringSlicePool = New(allocStringSlice, destroyStringSlice)
 
+func StringSlice() *Pool[*[]string] {
+	return stringSlicePool
+}
+
 func allocStringSlice() interface{} {
 	ret := make([]string, 0, 16)
 	return &ret
@@ -9,12 +13,4 @@ func allocStringSlice() interface{} {
 
 func destroyStringSlice(s *[]string) {
 	*s = (*s)[:0]
-}
-
-func GetStringSlice() *[]string {
-	return stringSlicePool.Get()
-}
-
-func ReleaseStringSlice(s *[]string) {
-	stringSlicePool.Put(s)
 }