--- a/internal/tokens/BUILD.bazel
+++ b/internal/tokens/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "tokens",
+    srcs = ["tokens.go"],
+    importpath = "github.com/lestrrat-go/jwx/v3/internal/tokens",
+    visibility = ["//:__subpackages__"],
+)
+
+alias(
+    name = "go_default_library",
+    actual = ":tokens",
+    visibility = ["//:__subpackages__"],
+)