--- a/internal/tokens/tokens.go
+++ b/internal/tokens/tokens.go
@@ -1,0 +1,12 @@
+package tokens
+
+const (
+	CloseCurlyBracket  = '}'
+	CloseSquareBracket = ']'
+	Colon              = ':'
+	Comma              = ','
+	DoubleQuote        = '"'
+	OpenCurlyBracket   = '{'
+	OpenSquareBracket  = '['
+	Period             = '.'
+)