--- a/jwa/BUILD.bazel
+++ b/jwa/BUILD.bazel
@@ -15,7 +15,7 @@ go_library(
     importpath = "github.com/lestrrat-go/jwx/v3/jwa",
     visibility = ["//visibility:public"],
     deps = [
-        "@com_github_lestrrat_go_option//:option",
+        "@com_github_lestrrat_go_option_v2//:option",
     ],
 )
 
@@ -33,7 +33,6 @@ go_test(
     deps = [
         ":jwa",
         "@com_github_stretchr_testify//require",
-        "@com_github_lestrrat_go_option//:option",
     ],
 )