--- a/jwa/compression_gen.go
+++ b/jwa/compression_gen.go
@@ -71,7 +71,9 @@ func NewCompressionAlgorithm(name string, options ...NewAlgorithmOption) Compres
 	for _, option := range options {
 		switch option.Ident() {
 		case identDeprecated{}:
-			deprecated = option.Value().(bool)
+			if err := option.Value(&deprecated); err != nil {
+				panic(fmt.Sprintf(`jwa: NewCompressionAlgorithm: %s`, err))
+			}
 		}
 	}
 	return CompressionAlgorithm{name: name, deprecated: deprecated}