--- a/jwa/key_type_gen.go
+++ b/jwa/key_type_gen.go
@@ -90,7 +90,9 @@ func NewKeyType(name string, options ...NewAlgorithmOption) KeyType {
 	for _, option := range options {
 		switch option.Ident() {
 		case identDeprecated{}:
-			deprecated = option.Value().(bool)
+			if err := option.Value(&deprecated); err != nil {
+				panic(fmt.Sprintf(`jwa: NewKeyType: %s`, err))
+			}
 		}
 	}
 	return KeyType{name: name, deprecated: deprecated}