--- a/jwa/signature_gen.go
+++ b/jwa/signature_gen.go
@@ -156,9 +156,13 @@ func NewSignatureAlgorithm(name string, options ...NewSignatureAlgorithmOption) 
 	for _, option := range options {
 		switch option.Ident() {
 		case identIsSymmetric{}:
-			isSymmetric = option.Value().(bool)
+			if err := option.Value(&isSymmetric); err != nil {
+				panic(fmt.Sprintf(`jwa: NewSignatureAlgorithm: %s`, err))
+			}
 		case identDeprecated{}:
-			deprecated = option.Value().(bool)
+			if err := option.Value(&deprecated); err != nil {
+				panic(fmt.Sprintf(`jwa: NewSignatureAlgorithm: %s`, err))
+			}
 		}
 	}
 	return SignatureAlgorithm{name: name, deprecated: deprecated, isSymmetric: isSymmetric}