--- a/jwe/BUILD.bazel
+++ b/jwe/BUILD.bazel
@@ -26,6 +26,7 @@ go_library(
         "//internal/json",
         "//internal/keyconv",
         "//internal/pool",
+        "//internal/tokens",
         "//jwa",
         "//jwe/internal/aescbc",
         "//jwe/internal/cipher",