--- a/jwe/io.go
+++ b/jwe/io.go
@@ -3,6 +3,7 @@
 package jwe
 
 import (
+	"fmt"
 	"io/fs"
 	"os"
 )
@@ -16,10 +17,12 @@ func (sysFS) Open(path string) (fs.File, error) {
 func ReadFile(path string, options ...ReadFileOption) (*Message, error) {
 
 	var srcFS fs.FS = sysFS{}
-	for _, option := range options {
-		switch option.Ident() {
+	for _, opt := range options {
+		switch opt.Ident() {
 		case identFS{}:
-			srcFS = option.Value().(fs.FS)
+			if err := opt.Value(&srcFS); err != nil {
+				return nil, fmt.Errorf("jwe.ReadFile: %s", err.Error())
+			}
 		}
 	}