This PR is actually one of the first times we did anything remotely close to optimization on this library. 
It shaves off about 10% of allocations, and memory pressure is generally better by about 15%~20% for many areas, including jws.Sign related operations.

This is definitely not a big change, but at least it's a start.

# Before

```
BenchmarkJWX_JWTSignHS256-24                      175449              6088 ns/op            4488 B/op        106 allocs/op
BenchmarkJWX_JWTSignRS256-24                        1471            789052 ns/op           13468 B/op        159 allocs/op
BenchmarkJWX_JWTParseHS256-24                     113784             10102 ns/op           13084 B/op        182 allocs/op
BenchmarkJWX_JWTParseRS256-24                      36052             32677 ns/op           15040 B/op        196 allocs/op
BenchmarkJWX_JWTParseWithValidation-24            110768             10435 ns/op           13085 B/op        182 allocs/op
BenchmarkJWX_CreateAndParseHS256-24                65100             17319 ns/op           17886 B/op        296 allocs/op
BenchmarkJWX_JWSSignHS256-24                      308347              3717 ns/op            3173 B/op         64 allocs/op
BenchmarkJWX_JWSParseHS256-24                     325042              3562 ns/op            3928 B/op         64 allocs/op
BenchmarkRSAPublicKeyMarshal-24                   506553              2196 ns/op            1581 B/op         11 allocs/op
BenchmarkRSAPrivateKeyMarshal-24                  148597              7814 ns/op            7650 B/op         32 allocs/op
BenchmarkECDSAPublicKeyMarshal-24                 757569              1657 ns/op             979 B/op         14 allocs/op
BenchmarkECDSAPrivateKeyMarshal-24                493706              2150 ns/op            1389 B/op         17 allocs/op
BenchmarkSymmetricKeyMarshal-24                  1358782               881.1 ns/op           489 B/op          8 allocs/op
BenchmarkOKPEd25519PublicKeyMarshal-24           1221278               971.2 ns/op           385 B/op         11 allocs/op
BenchmarkOKPEd25519PrivateKeyMarshal-24           936276              1317 ns/op             634 B/op         14 allocs/op
BenchmarkOKPX25519PublicKeyMarshal-24            1217532               983.4 ns/op           377 B/op         11 allocs/op
BenchmarkOKPX25519PrivateKeyMarshal-24            837657              1300 ns/op             594 B/op         14 allocs/op
BenchmarkSetMultipleKeysMarshal/WithoutCustomFields-24             57438             20201 ns/op           12476 B/op         70 allocs/op
BenchmarkSetMultipleKeysMarshal/WithCustomFields-24                40058             29342 ns/op           15510 B/op        105 allocs/op
```

# After

```
BenchmarkJWX_JWTSignHS256-24                      198655              5725 ns/op            3020 B/op         82 allocs/op
BenchmarkJWX_JWTSignRS256-24                        1491            793797 ns/op           10392 B/op        133 allocs/op
BenchmarkJWX_JWTParseHS256-24                     119838             10450 ns/op           12959 B/op        190 allocs/op
BenchmarkJWX_JWTParseRS256-24                      36111             34375 ns/op           14967 B/op        205 allocs/op
BenchmarkJWX_JWTParseWithValidation-24            110686             10274 ns/op           12959 B/op        190 allocs/op
BenchmarkJWX_CreateAndParseHS256-24                66962             17422 ns/op           16421 B/op        280 allocs/op
BenchmarkJWX_JWSSignHS256-24                      360300              3226 ns/op            1879 B/op         41 allocs/op
BenchmarkJWX_JWSParseHS256-24                     281103              3927 ns/op            4072 B/op         72 allocs/op
BenchmarkRSAPublicKeyMarshal-24                   577527              2069 ns/op            1584 B/op         11 allocs/op
BenchmarkRSAPrivateKeyMarshal-24                  153822              7670 ns/op            7048 B/op         29 allocs/op
BenchmarkECDSAPublicKeyMarshal-24                 633177              1739 ns/op             982 B/op         14 allocs/op
BenchmarkECDSAPrivateKeyMarshal-24                514675              2190 ns/op            1392 B/op         17 allocs/op
BenchmarkSymmetricKeyMarshal-24                  1265098               911.0 ns/op           490 B/op          8 allocs/op
BenchmarkOKPEd25519PublicKeyMarshal-24           1000000              1013 ns/op             386 B/op         11 allocs/op
BenchmarkOKPEd25519PrivateKeyMarshal-24           782100              1411 ns/op             635 B/op         14 allocs/op
BenchmarkOKPX25519PublicKeyMarshal-24            1000000              1025 ns/op             378 B/op         11 allocs/op
BenchmarkOKPX25519PrivateKeyMarshal-24            828104              1331 ns/op             595 B/op         14 allocs/op
BenchmarkSetMultipleKeysMarshal/WithoutCustomFields-24             57260             20476 ns/op           11887 B/op         67 allocs/op
BenchmarkSetMultipleKeysMarshal/WithCustomFields-24                41182             29498 ns/op           14452 B/op         98 allocs/op
```